pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pet-id"
version = "0.1.0"
description = "Pet identity system: multi-view registration + attribution (bootstrap)"
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "tenacity>=8.2",
    "python-json-logger>=2.0",
    "opencv-python-headless>=4.9",
    "pillow>=10.0",
]

[project.optional-dependencies]
detector = ["ultralytics>=8.3"]
reid     = ["torchreid>=0.2.5", "torch>=2.1", "torchvision>=0.16"]
pose     = ["mmpose>=1.3", "mmcv>=2.1", "mmengine>=0.10"]
narrative= ["transformers>=4.45", "accelerate>=0.30", "torch>=2.1", "qwen-vl-utils"]
tracker  = ["lap>=0.4", "scipy>=1.11", "boxmot>=10.0"]
all      = ["pet-id[detector,reid,pose,narrative,tracker]"]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["purrai_core", "purrai_core.*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = ["ultralytics", "ultralytics.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["boxmot", "boxmot.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["torchreid", "torchreid.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["mmpose", "mmpose.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["mmcv", "mmcv.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["mmengine", "mmengine.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["transformers", "transformers.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["PIL", "PIL.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["torch", "torch.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["pythonjsonlogger", "pythonjsonlogger.*"]
ignore_missing_imports = true
follow_imports = "skip"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
markers = [
    "gpu: tests requiring CUDA",
    "slow: slow tests (>5s)",
]
